+++ src/init.h
@@ -14,6 +14,8 @@ namespace forge
   {
     bool github = false;
     std::optional<std::string> library_type;
+    std::optional<std::string> initial_version;
+    std::optional<std::filesystem::path> version_header_path;
   };
 
   int adopt_project(const std::filesystem::path& project_directory,